Renamed crate to devices_test_helpers
