+++ vm/devices/devices_test_helpers/Cargo.toml
@@ -1,0 +1,22 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+[package]
+name = "devices_test_helpers"
+version = "0.1.0"
+edition.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+chipset_device.workspace = true
+guestmem.workspace = true
+inspect.workspace = true
+parking_lot.workspace = true
+pci_core.workspace = true
+safeatomic.workspace = true
+tracing.workspace = true
+user_driver.workspace = true
+
+[lints]
+workspace = true
