+++ vm/devices/devices_test_helpers/src/emulated_test_infra.rs
