+++ vm/devices/net/mana_driver/src/tests.rs
@@ -9,10 +9,9 @@ use crate::bnic_driver::RxConfig;
 use crate::bnic_driver::WqConfig;
 use crate::gdma_driver::GdmaDriver;
 use crate::mana::ResourceArena;
-use dma_guest_memory_access_wrapper::emulated::DeviceSharedMemory;
-use dma_guest_memory_access_wrapper::emulated::EmulatedDevice;
-use dma_guest_memory_access_wrapper::emulated::EmulatedDmaAllocator;
-use dma_guest_memory_access_wrapper::memory::PAGE_SIZE;
+use devices_test_helpers::emulated_test_infra::DeviceSharedMemory;
+use devices_test_helpers::emulated_test_infra::EmulatedDevice;
+use devices_test_helpers::emulated_test_infra::EmulatedDmaAllocator;
 use chipset_device::mmio::ExternallyManagedMmioIntercepts;
 use gdma::VportConfig;
 use gdma_defs::GdmaDevType;
@@ -23,6 +22,7 @@ use pal_async::async_test;
 use pci_core::msi::MsiInterruptSet;
 use std::sync::Arc;
 use test_with_tracing::test;
+use user_driver::memory::PAGE_SIZE;
 use user_driver::DeviceBacking;
 use vmcore::vm_task::SingleDriverBackend;
 use vmcore::vm_task::VmTaskDriverSource;
