+++ vm/devices/net/net_mana/Cargo.toml
@@ -27,7 +27,7 @@ zerocopy.workspace = true
 
 [dev-dependencies]
 chipset_device.workspace = true
-dma_guest_memory_access_wrapper.workspace = true
+devices_test_helpers.workspace = true
 gdma.workspace = true
 pci_core.workspace = true
 test_with_tracing.workspace = true
