+++ vm/devices/storage/disk_nvme/nvme_driver/Cargo.toml
@@ -32,7 +32,7 @@ zerocopy.workspace = true
 [dev-dependencies]
 chipset_device.workspace = true
 disklayer_ram.workspace = true
-dma_guest_memory_access_wrapper.workspace = true
+devices_test_helpers.workspace = true
 guid.workspace = true
 nvme.workspace = true
 pci_core.workspace = true
