+++ vm/devices/storage/disk_nvme/nvme_driver/fuzz/fuzz_emulated_device.rs
@@ -11,11 +11,11 @@ use crate::arbitrary_data;
 
 use chipset_device::mmio::MmioIntercept;
 use chipset_device::pci::PciConfigSpace;
-use dma_guest_memory_access_wrapper::emulated::DeviceSharedMemory;
-use dma_guest_memory_access_wrapper::emulated::EmulatedDevice;
-use dma_guest_memory_access_wrapper::emulated::EmulatedDmaAllocator;
-use dma_guest_memory_access_wrapper::emulated::Mapping;
-use dma_guest_memory_access_wrapper::interrupt::DeviceInterrupt;
+use devices_test_helpers::emulated_test_infra::DeviceSharedMemory;
+use devices_test_helpers::emulated_test_infra::EmulatedDevice;
+use devices_test_helpers::emulated_test_infra::EmulatedDmaAllocator;
+use devices_test_helpers::emulated_test_infra::Mapping;
+use devices_test_helpers::interrupt::DeviceInterrupt;
 use inspect::Inspect;
 use inspect::InspectMut;
 use pci_core::msi::MsiInterruptSet;
