+++ vm/devices/storage/disk_nvme/nvme_driver/fuzz/fuzz_nvme_driver.rs
@@ -7,7 +7,7 @@ use crate::fuzz_emulated_device::FuzzEmulatedDevice;
 
 use arbitrary::Arbitrary;
 use chipset_device::mmio::ExternallyManagedMmioIntercepts;
-use dma_guest_memory_access_wrapper::emulated::DeviceSharedMemory;
+use devices_test_helpers::emulated_test_infra::DeviceSharedMemory;
 use guestmem::GuestMemory;
 use guid::Guid;
 use nvme::NvmeController;
