+++ vm/devices/storage/disk_nvme/nvme_driver/src/tests.rs
@@ -5,9 +5,9 @@ use crate::NvmeDriver;
 use chipset_device::mmio::ExternallyManagedMmioIntercepts;
 use chipset_device::mmio::MmioIntercept;
 use chipset_device::pci::PciConfigSpace;
-use dma_guest_memory_access_wrapper::emulated::EmulatedDevice;
-use dma_guest_memory_access_wrapper::emulated::Mapping;
-use dma_guest_memory_access_wrapper::emulated::GuestMemoryAccessWrapper;
+use devices_test_helpers::emulated_test_infra::EmulatedDevice;
+use devices_test_helpers::emulated_test_infra::Mapping;
+use devices_test_helpers::emulated_test_infra::GuestMemoryAccessWrapper;
 use guestmem::GuestMemory;
 use guid::Guid;
 use inspect::Inspect;
